fix: use new execution engine
